Barebones definition of FDW
This is a barebones definition of a FDW. It provides only the callbacks
required for a read-only FDW, but does not implement the actual hooks
to call the java code.

The header also calls out the three functions we're mostly likely to
need to define. They are:

 - create_foreignscan_path
 - make_foreignscan

and a third which creates the resulting tuples. I didn't pursue that
further since it's a solved problem - there is already a 'Tuple.h' type.

(Note: it *might* also require explainForeignScan and even
explainForeignModify - I've seen conflicting information.)